+++ lib/nekonote/request.rb
@@ -1,11 +1,5 @@
 module Nekonote
     class Request
-        STRING  = 'string'
-        INTEGER = 'int'
-        ARRAY   = 'array'
-        FLOAT   = 'float'
-        BOOL    = 'bool'
-
         # accessor
         attr_reader :path,
                     :uri,
@@ -13,43 +7,36 @@ module Nekonote
                     :payload
 
         # @param hash env
-        # @param string|nil restricted
-        def initialize(env, restricted = nil)
+        def initialize(env)
             # set properties
             @path   = env['REQUEST_PATH']
             @uri    = env['REQUEST_URI']
             @method = env['REQUEST_METHOD']
-            @accepted_list = get_accepted_list restricted
 
             # query string
-            @query_string     = {}
-            @query_string_raw = {}
+            @query_string = {}
 
             # POST data
-            @post_data     = {}
-            @post_data_raw = {}
+            @post_data = {}
 
             # URL path parameters
-            @path_params     = {}
-            @path_params_raw = {}
+            @path_params = {} # this will be set from handler
 
             # request body
             @payload = nil
 
             # set query string
             if env['QUERY_STRING'].is_a? String
-                @query_string_raw = get_param_maps env['QUERY_STRING']
-                @query_string     = get_sanitized @query_string_raw
+                @query_string = get_param_maps env['QUERY_STRING']
             end
 
             # set POST data
             if env['rack.input'].is_a? StringIO
-                @post_data_raw = get_param_maps env['rack.input'].gets
-                @post_data     = get_sanitized @post_data_raw
+                @post_data = get_param_maps env['rack.input'].gets
             end
         end
 
-        # Returns sanitized query string
+        # Returns query string
         # @params string|symbol name
         # @return mixed
         public
@@ -57,15 +44,7 @@ module Nekonote
             return get_value name, @query_string
         end
 
-        # Returns raw query string
-        # @params string|symbol name
-        # @return mixed
-        public
-        def query_string_raw(name = nil)
-            return get_value name, @query_string_raw
-        end
-
-        # Returns sanitized POST data
+        # Returns POST data
         # @params string|symbol name
         # @return mixed
         public
@@ -73,28 +52,12 @@ module Nekonote
             return get_value name, @post_data
         end
 
-        # Returns raw POST data
-        # @params string|symbol name
-        # @return mixed
-        public
-        def post_data_raw(name = nil)
-            return get_value name, @post_data_raw
-        end
-
-        # Returns sanitized URL parameters
+        # Returns URL parameters
         # @params string|symbol name
         # @return mixed
         public
         def path_params(name = nil)
             return get_value name, @path_params
-        end
-
-        # Returns raw URL parameters
-        # @params string|symbol name
-        # @return mixed
-        public
-        def path_params_raw(name = nil)
-            return get_value name, @path_params_raw
         end
 
         # Returns sanitized parameters
@@ -107,22 +70,6 @@ module Nekonote
                 maps = @query_string
             when 'POST'
                 maps = @post_data
-            else
-                maps = {}
-            end
-            return get_value name, maps
-        end
-
-        # Returns raw parameters
-        # @params string|symbol name
-        # @return mixed
-        public
-        def params_raw(name = nil)
-            case @method
-            when 'GET'
-                maps = @query_string_raw
-            when 'POST'
-                maps = @post_data_raw
             else
                 maps = {}
             end
@@ -154,8 +101,7 @@ module Nekonote
             return nil if !map.is_a? Hash
 
             # set url params
-            @path_params_raw = map
-            @path_params     = get_sanitized @path_params_raw
+            @path_params = map
         end
 
         # @param mixed payload
@@ -193,27 +139,6 @@ module Nekonote
             end
         end
 
-        # Returns the accepable name list
-        # @param string|nil restricted
-        # @return array
-        private
-        def get_accepted_list(restricted)
-            accepted_list = []
-            if restricted.is_a? String
-                restricted.split(',').each do |data|
-                    data.strip!
-                    pair = data.split('=')
-                    if pair.size != 2
-                        raise Error, Error::MSG_INVALID_FIELD%['params', Preference.instance.path_route_yml]
-                    end
-                    # add field name and hash as Hash into array
-                    accepted_list << {pair[0] => pair[1]}
-                end
-            end
-
-            return accepted_list
-        end
-
         # @param string str query string
         # @return hash
         private
@@ -232,64 +157,6 @@ module Nekonote
             end
 
             return maps
-        end
-
-        # Removing waste parameter and converting to the exected type
-        # @param hash maps
-        # @param bool decode
-        # @return hash
-        private
-        def get_sanitized(maps, decode = false)
-            return maps if @accepted_list.size == 0
-
-            sanitized_maps = {}
-            @accepted_list.each do |field|
-                field.each_pair do |name, type|
-                    converted = get_expected type, maps[name]
-
-                    # URL decode if necessary
-                    if decode || converted.is_a?(String)
-                        converted = URI.decode_www_form_component converted
-                    end
-
-                    sanitized_maps[name] = converted
-                end
-            end
-
-            return sanitized_maps
-        end
-
-        # @param string type type expected
-        # @param mixed value real value
-        # @param string converted value by the expected type
-        private
-        def get_expected(type, value)
-            return nil if value == nil
-
-            v = nil
-            begin
-                case type
-                when INTEGER
-                    v = value.to_i
-                when STRING
-                    v = value.to_s
-                when ARRAY
-                    v = value.to_s.split(',')
-                    v.map! { |val| val.strip }
-                when FLOAT
-                    v = value.to_f
-                when BOOL
-                    if value == nil || value == false || value == 'false' || value == 0 || value == '0' || value == 'no' || value == ''
-                        v = false
-                    else
-                        v = true
-                    end
-                end
-            rescue
-                # value is nil if failed to convert value to expected type
-            end
-
-            return v
         end
     end
 end