+++ lib/nekonote/view.rb
@@ -14,7 +14,6 @@ module Nekonote
                       :info_path,
                       :info_exec_method,
                       :info_allow_method,
-                      :info_params,
                       :info_content_type,
                       :info_template,
                       :info_layout,
@@ -260,7 +259,6 @@ module Nekonote
             @info_path            = info[Preference::FIELD_ROUTE_PATH]
             @info_exec_method     = info[Preference::FIELD_ROUTE_EXEC_METHOD]
             @info_allow_method    = info[Preference::FIELD_ROUTE_ALLOW_METHOD]
-            @info_params          = info[Preference::FIELD_ROUTE_PARAMS]
             @info_content_type    = info[Preference::FIELD_ROUTE_CONTENT_TYPE]
             @info_template        = info[Preference::FIELD_ROUTE_TEMPLATE]
             @info_layout          = info[Preference::FIELD_ROUTE_LAYOUT]